+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-31 16:50:35
-// Commit: test(theme): implement widget integration tests for theme
+// Updated at: 2026-08-31 18:42:21
+// Commit: feat(theme): add bi-directional event bus handler for theme
 
 class themeComponentState {
-  final String commitRef = '4085';
-  final String lastMessage = 'test(theme): implement widget integration tests for theme';
-  final DateTime timestamp = DateTime.parse('2026-08-31T16:50:35+05:00');
+  final String commitRef = '4089';
+  final String lastMessage = 'feat(theme): add bi-directional event bus handler for theme';
+  final DateTime timestamp = DateTime.parse('2026-08-31T18:42:21+05:00');
 }